--- src/AboutPage/about.scss
@@ -1,19 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-
-.about {
-    max-width: 1440px;
-    display: flex;
-    justify-content: space-between;
-    font-size: 14px;
-    line-height: 20px;
-    margin:20px 135px 0 135px;
-
-    .about_text {
-        max-width: 579px;
-        padding-right: 5px;
-    }
-}
