+++ src/Components/Footer/Footer.scss
@@ -1,6 +1,6 @@
-@import '../commonStyles/reset.scss';
-@import '../commonStyles/variables.scss';
-@import '../commonStyles/font-family.scss';
+@import './../../commonStyles/reset.scss';
+@import './../../commonStyles/variables.scss';
+@import './../../commonStyles/font-family.scss';
 
 .footer_main {
     display: flex;
