+++ src/commonStyles/font-family.scss
@@ -7,3 +7,13 @@
     font-family: 'PoppinsBold';
     src: url('../assets/fonts/Poppins/Poppins-Bold.ttf');
 }
+
+@font-face {
+    font-family: 'Baloo2Regular';
+    src: url('../assets/fonts/Baloo2/Baloo2-Regular.ttf');
+}
+
+@font-face {
+    font-family: 'Baloo2Bold';
+    src: url('../assets/fonts/Baloo2/Baloo2-Bold.ttf');
+}
